fix: Ghana page metadata — split into server wrapper and client component
